Added retaining personal information functionality for booking multiple appointments

+++ js/scripts.js
@@ -1,3 +1,9 @@
 function onLoad() {
-    localStorage.clear();
+    console.log(localStorage.getItem("fullname"));
+    console.log(localStorage.getItem("email"));
+    console.log(localStorage.getItem("domain"));
+    var url = window.location.href;
+    if (localStorage.getItem("fullname") == null || url.includes(index.html)) {
+        localStorage.clear();
+    }
 }
